:hammer: Fix: table 필드 수정으로 인한 엔티티 수정

+++ back/Dream-Hi/src/main/java/com/elephant/dreamhi/model/entity/ActorProfile.java
@@ -48,7 +48,7 @@ public class ActorProfile {
     private Integer age;
     @Enumerated(EnumType.STRING)
     private Gender gender;
-    private Double height;
+    private Integer height;
     @Column(length = 500)
     private String description;
     @Column(nullable = false)
